requirements.txt
streamlit
pandas
pytest
scikit-learn
shap
lime
requests
apify-client
python-dotenv